# util: improve inspect performance