--- a/lib/internal/util.js
+++ b/lib/internal/util.js
@@ -252,43 +252,6 @@ function getSystemErrorName(err) {
   return entry ? entry[0] : `Unknown system error ${err}`;
 }
 
-// getConstructorOf is wrapped into this to save iterations
-function getIdentificationOf(obj) {
-  const original = obj;
-  let constructor;
-  let tag;
-
-  while (obj) {
-    if (constructor === undefined) {
-      const desc = Object.getOwnPropertyDescriptor(obj, 'constructor');
-      if (desc !== undefined &&
-          typeof desc.value === 'function' &&
-          desc.value.name !== '')
-        constructor = desc.value.name;
-    }
-
-    if (tag === undefined) {
-      const desc = Object.getOwnPropertyDescriptor(obj, Symbol.toStringTag);
-      if (desc !== undefined) {
-        if (typeof desc.value === 'string') {
-          tag = desc.value;
-        } else if (desc.get !== undefined) {
-          tag = desc.get.call(original);
-          if (typeof tag !== 'string')
-            tag = undefined;
-        }
-      }
-    }
-
-    if (constructor !== undefined && tag !== undefined)
-      break;
-
-    obj = Object.getPrototypeOf(obj);
-  }
-
-  return { constructor, tag };
-}
-
 const kCustomPromisifiedSymbol = Symbol('util.promisify.custom');
 const kCustomPromisifyArgsSymbol = Symbol('customPromisifyArgs');
 
@@ -431,7 +394,6 @@ module.exports = {
   filterDuplicateStrings,
   getConstructorOf,
   getSystemErrorName,
-  getIdentificationOf,
   isError,
   isInsideNodeModules,
   join,