--- a/lib/util.js
+++ b/lib/util.js
@@ -574,6 +574,11 @@ function formatValue(ctx, value, recurseTimes, ln) {
         braces[0] = '[Arguments] {';
         if (keyLength === 0)
           return '[Arguments] {}';
+      } else if (tag !== '') {
+        braces[0] = `${getPrefix(constructor, tag)}{`;
+        if (keyLength === 0) {
+          return `${braces[0]}}`;
+        }
       } else if (keyLength === 0) {
         return '{}';
       }