This improves a slow code part in `util.inspect` by directly
retrieving the `Symbol.toStringTag` and by optimizing some code
paths.

I benchmarked this with #20007 applied.

```js
'use strict';

const util = require('util');

let i = 0;
const obj = {};

while (i++ < 100000) {
  obj[i] = {
    a: new Set([1, {}]),
    b: [],
    c: [1, 2],
    d: true
  };
}

for (const k of Object.keys(obj)) {
  obj[k].obj = obj;
}

console.time('run');
util.inspect(obj);
console.timeEnd('run');

// node$ ./slow t.js
// run: 692.284ms
// run: 681.384ms
// run: 685.455ms

// node$ ./fast t.js
// run: 645.635ms
// run: 640.754ms
// run: 648.319ms
```

<!--
Thank you for your pull request. Please provide a description above and review
the requirements below.

Bug fixes and new features should include tests and possibly benchmarks.

Contributors guide: https://github.com/nodejs/node/blob/master/CONTRIBUTING.md
-->

##### Checklist
<!-- Remove items that do not apply. For completed items, change [ ] to [x]. -->

- [x] `make -j4 test` (UNIX), or `vcbuild test` (Windows) passes
- [x] tests and/or benchmarks are included
- [x] documentation is changed or added
- [x] commit message follows [commit guidelines](https://github.com/nodejs/node/blob/master/doc/guides/contributing/pull-requests.md#commit-message-guidelines)