@jasnell 2018-04-13
Can you post the benchmark results?