@BridgeAR 2018-04-13
@jasnell I posted them in the description below the code?