@jasnell 2018-04-13
oh, missed that. Is there not an existing benchmark script that works with compare.js? Seeing those results would be more reliable I think.