@BridgeAR 2018-04-16
Rebased

New CI https://ci.nodejs.org/job/node-test-pull-request/14332/